+++ lib-multisrc/raijinscans/src/eu/kanade/tachiyomi/multisrc/raijinscans/RaijinScansDto.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.extension.fr.raijinscans
+package eu.kanade.tachiyomi.multisrc.raijinscans
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -14,14 +14,4 @@ class LatestUpdatesDataDto(
     @SerialName("manga_html") val mangaHtml: String,
     @SerialName("current_page") val currentPage: Int,
     @SerialName("total_pages") val totalPages: Int,
-)
-
-@Serializable
-class ImageResponseDto(
-    val data: ImageResponseDataDto,
-)
-
-@Serializable
-class ImageResponseDataDto(
-    val u: List<String>,
 )
