+++ src/fr/pornhwaio/build.gradle
@@ -1,0 +1,10 @@
+ext {
+    extName = 'Pornhwa.io'
+    extClass = '.PornhwaIo'
+    themePkg = 'raijinscans'
+    baseUrl = 'https://pornhwa.io'
+    overrideVersionCode = 0
+    isNsfw = true
+}
+
+apply from: "$rootDir/common.gradle"
